requirements.txt
python-dotenv
groq
tavily-python
requests
firecrawl-py
firecrawl